update user class so the boardids are long instead of int
the user class used to contain integers for the board ids, now they are longs because this is what the board ids use aswell

+++ commons/src/main/java/commons/User.java
@@ -5,7 +5,7 @@ import java.util.List;
 public class User {
 
     private String name;
-    private List<Integer> boardIds;
+    private List<Long> boardIds;
 
     public User(String name) {
         this.name = name;
@@ -27,7 +27,7 @@ public class User {
         this.boardIds = boardIds;
     }
 
-    public boolean containsBoardId(int id) {
+    public boolean containsBoardId(long id) {
         return boardIds.contains(id);
     }
 }
